Update peer-utils.js
I'm proposing this file change to fix a problem that I've been having. 

I have a peculiar setup, in my setup getPeerDependency('cucumber') should return something like this:

 C:\MyProject\node_modules\my-private-module\node_modules\cucumber\lib\cucumber.js

Instead the function is returning

 C:\MyProject\node_modules\cucumber\lib\cucumber.js

This is because the regexp truncates packagePath at the first occurrence of node_modules that it finds, instead I think it should truncate starting from the last occurrence that it finds.

Let me know if this doesn't make sense, if so I will try and reproduce the problem.